+++ HFramework/src/Runtime/Events/EventHandlers/NormalCountEventHandler.cs
@@ -12,8 +12,13 @@ namespace HFramework.Events.EventHandlers
 
 		internal void Init() {
 			// Only penetration of ass/vagina are counted as normal sex
+#pragma warning disable CS0618 // Type or member is obsolete
 			SexEvents.OnPenetrateVagina.Triggered += this.OnPenetrationStart;
 			SexEvents.OnPenetrateAss.Triggered += this.OnPenetrationStart;
+#pragma warning restore CS0618 // Type or member is obsolete
+
+			SexEvents.OnInteractPenis2Vagina.Triggered += this.OnPenetrationStart;
+			SexEvents.OnInteractPenis2Ass.Triggered += this.OnPenetrationStart;
 		}
 
 		private void OnPenetrationStart(object sender, FromToEventArgs e) {