+++ HFramework/src/Runtime/Events/EventHandlers/RapeCountEventHandler.cs
@@ -11,6 +11,7 @@ namespace HFramework.Events.EventHandlers
 		internal static RapeCountEventHandler Instance { get; private set; } = new RapeCountEventHandler();
 
 		internal void Init() {
+#pragma warning disable CS0618 // Type or member is obsolete
 			SexEvents.OnPenetrateVagina.Triggered += this.OnActStart;
 			SexEvents.OnPenetrateAss.Triggered += this.OnActStart;
 			SexEvents.OnPerformHandJob.Triggered += this.OnActStart;
@@ -18,6 +19,14 @@ namespace HFramework.Events.EventHandlers
 			SexEvents.OnPerformScissor.Triggered += this.OnActStart;
 			SexEvents.OnPenetrateMouth.Triggered += this.OnActStart;
 			SexEvents.OnLickVagina.Triggered += this.OnActStart;
+#pragma warning restore CS0618 // Type or member is obsolete
+			SexEvents.OnInteractPenis2Vagina.Triggered += this.OnActStart;
+			SexEvents.OnInteractPenis2Ass.Triggered += this.OnActStart;
+			SexEvents.OnInteractPenis2Mouth.Triggered += this.OnActStart;
+			SexEvents.OnInteractPenis2Hand.Triggered += this.OnActStart;
+			SexEvents.OnInteractPenis2Tits.Triggered += this.OnActStart;
+			SexEvents.OnInteractVagina2Vagina.Triggered += this.OnActStart;
+			SexEvents.OnInteractTongue2Vagina.Triggered += this.OnActStart;
 		}
 
 		private void OnActStart(object sender, FromToEventArgs e) {